+++ zellij-utils/src/consts.rs
@@ -1,5 +1,6 @@
 //! Zellij program-wide constants.
 
+use crate::input::theme::Themes;
 use directories_next::ProjectDirs;
 use lazy_static::lazy_static;
 use once_cell::sync::OnceCell;
@@ -28,6 +29,19 @@ lazy_static! {
     pub static ref ZELLIJ_PROJ_DIR: ProjectDirs =
         ProjectDirs::from("org", "Zellij Contributors", "Zellij").unwrap();
     pub static ref ZELLIJ_CACHE_DIR: PathBuf = ZELLIJ_PROJ_DIR.cache_dir().to_path_buf();
+    pub static ref ZELLIJ_DEFAULT_THEMES: Themes = {
+        let mut default_themes = Themes::default();
+
+        let path = PathBuf::from(concat!(env!("CARGO_MANIFEST_DIR"), "/assets/themes"));
+        match Themes::from_dir(path) {
+            Ok(themes) => {
+                default_themes = default_themes.merge(themes);
+            },
+            Err(_) => {},
+        }
+
+        default_themes
+    };
 }
 
 pub const FEATURES: &[&str] = &[