+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * feat: allow excluding tabs from tab sync in layouts (https://github.com/zellij-org/zellij/pull/2314)
 * feat: support default cwd (https://github.com/zellij-org/zellij/pull/2290)
 * feat: cli action to reload plugins at runtime for easier plugin development (https://github.com/zellij-org/zellij/pull/2372)
+* docs(architecture): update architecture docs (https://github.com/zellij-org/zellij/pull/2371)
+* feat(themes): add nightfox themes (https://github.com/zellij-org/zellij/pull/2384)
+* feat: provide default themes (https://github.com/zellij-org/zellij/pull/2307)
 
 ## [0.36.0] - 2023-04-13
 * fix: when moving pane focus off screen edge to the next tab, the pane on the screen edge is now focused (https://github.com/zellij-org/zellij/pull/2293)